AST: Split off TypeSubstitution.cpp from Type.cpp
