+++ lib/AST/CMakeLists.txt
@@ -117,6 +117,7 @@ add_swift_host_library(swiftAST STATIC
   TypeJoinMeet.cpp
   TypeRefinementContext.cpp
   TypeRepr.cpp
+  TypeSubstitution.cpp
   TypeWalker.cpp
   UnqualifiedLookup.cpp
   USRGeneration.cpp
